+++ plugins/binary/crates/plugin/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod behaviour;
-pub mod plugin;
-pub mod providers;
-pub mod web_resource_provider;