+++ plugins/file/crates/model/src/component/mod.rs
@@ -1,5 +1,0 @@
-pub use file::*;
-pub use fs_notify::*;
-
-pub mod file;
-pub mod fs_notify;