+++ plugins/file/crates/plugin/src/providers.rs
@@ -1,5 +1,0 @@
-use reactive_graph_plugin_api::prelude::providers::*;
-
-#[derive(TypeProvider, Component)]
-#[type_provider(tys = "Components", path = "types/components")]
-pub struct ValueComponentsProvider {}