+++ plugins/system-environment/crates/model/src/entity/mod.rs
@@ -1,3 +1,0 @@
-pub use system_env::*;
-
-pub mod system_env;