+++ plugins/system-environment/crates/plugin/src/lib.rs
@@ -1,7 +1,0 @@
-#![allow(clippy::map_entry)]
-
-use reactive_graph_model_system_environment as model_system_environment;
-
-pub mod factory;
-pub mod plugin;
-pub mod providers;